Drawables are now updated

+++ Java Game Engine/src/engine/GameLogic.java
@@ -4,5 +4,5 @@ public abstract class GameLogic {
 
     public abstract void init();
 
-    public abstract void update(long time);
+    public abstract void update(double time);
 }
